Split fields up and down testing, fixed visibility
